+++ app/components/index.ts
@@ -3,4 +3,5 @@ export { default as Footer } from './Footer';
 export { default as LEDMeterBar } from './LEDMeterBar';
 export { default as PluginCard } from './PluginCard';
 export { default as PluginList } from './PluginList';
+export { ErrorBoundary } from './ErrorBoundary';
 export type { PluginData } from './PluginCard';